+++ changelog.md
@@ -4,6 +4,17 @@
 ---
 
 
+## [1.0.1] - 2018-11-14
+
+
+### Changed
+
+- Updated starter app to `@architect/architect` 4.3.8
+- Updated starter app bundled tests
+
+---
+
+
 ## [1.0.0] - 2018-11-04
 
 
@@ -13,6 +24,7 @@
 
 
 ### Changed
+
 - Starter app now uses `@architect/architect` 4.x (specifically, `4.1.2`)
 - `.arc` file and test suite updated for to use new `@http` functions
 - npm does not allow the publishing of `package-lock.json` files, but it does allow the publishing of `npm-shrinkwrap.json` files, so those files are renamed here (but will renamed back again by Begin when provisioned to customers)